requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-docx>=1.1.0
habanero>=1.2.0
citeproc-py>=0.6.0
jinja2>=3.1.0
pydantic>=2.5.0
aiofiles>=23.2.0
httpx>=0.26.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
